Hack around strange change in ssh behaviour.
Before 16.04 as host OS, we would start a VM and then soon
after ssh to it.  This ssh would block while the VM got started
then proceed normally.

Now we get a "connection closed by remote host", so we have to
ignore this and keep trying until the VM is up.