Don't build a tree that is being packaged; the package() method must do it (if required).
This is to make Linux builds more efficient; previously cdist
would build() everything (including the whole of DoM) and then
most linux package() methods would end up building it again
(making the build in build() redundant).

Now the cscript in DoM can call build() if necessary.

I'm a little bit worried that the environment is no longer preserved
during the final build(), but I'm unsure if this is necessary
anyway.